Add custom fixture node.
Needs customFixtures.json inside data folder. With an array of "Name" : "config" (both as strings). With name being the name of the node to be created and config the configuration in the form of "{paramname1}:{default value}, {paramname2}:{default value}:
To create a 16 bit parameter you need to append "//16" to the name.
The default value should go from 0 to 255 for default parameters, and from 0 to 65535 for 16 bit parameters.